Completed heuristic AI with UI controls.
